Move TagNoteMap under note package + init test files
