+++ src/test/java/seedu/address/model/note/NoteTest.java
@@ -1,0 +1,18 @@
+package seedu.address.model.note;
+
+import org.junit.jupiter.api.Test;
+
+class NoteTest {
+
+    @Test
+    void testEquals() {
+    }
+
+    @Test
+    void testToString() {
+    }
+
+    @Test
+    void testHashCode() {
+    }
+}
