+++ src/test/java/seedu/address/model/note/TagNoteMapTest.java
@@ -1,0 +1,26 @@
+package seedu.address.model.note;
+
+import org.junit.jupiter.api.Test;
+
+class TagNoteMapTest {
+
+    @Test
+    void initializeLists() {
+    }
+
+    @Test
+    void getTagsForNote() {
+    }
+
+    @Test
+    void getNotesForTag() {
+    }
+
+    @Test
+    void updateTagsForNote() {
+    }
+
+    @Test
+    void updateNotesForTag() {
+    }
+}
